==> cosmx/analysis/3_neighborhoods_230207.Rmd <==
---
title: "Neighborhood analysis"
author: "Saskia Freytag"
date: "2022-12-21"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(vegan)
library(Seurat)
library(tidyverse)
library(parallel)
library(pals)
library(pheatmap)
library(scater)
library(scran)
library(RColorBrewer)

source("../scripts/tiling.R")
```

## Load data

```{r}
sce <- readRDS("../data/processed/sce_jj98_final.rds")
anno <- readRDS("../data/processed/annotation_all.rds")
sce$Anno <- anno$Annotation
sce$Barcode <- sce$id
sce$CenterX_local_px <- anno$CenterX_local_px
sce$CenterY_local_px <- anno$CenterY_local_px
```



```{r}
df_polygons <- read.csv("../data/raw/SMI-0151_SarahBest_WEHI/5 Raw data/JJ98/JJ98-polygons.csv")
df_polygons$Barcode <- paste0("JJ98_", df_polygons$cellID, "_", df_polygons$fov)

# put annotation 
df_polygons$Anno <- sce$Anno[match(df_polygons$Barcode, sce$Barcode)]
df_polygons <- df_polygons[!df_polygons$Anno=="Not available",]
df_polygons$Anno <- factor(df_polygons$Anno)

```

## Employ tiling 

To ensure tiling is not influenced by cells that were excluded from analysis, I remove them.

```{r}
df_polygons <- df_polygons[!df_polygons$Anno=="Not available",]
df_polygons$Anno <- factor(df_polygons$Anno)
```

```{r}
set.seed(12)
df_final <- find_niches(df_polygons, tile_height=500, tile_shift=100, image_size=4400, num_clusters=5,
            cores=10, prob=FALSE)


cols <- brewer.set1(9)
names(cols) <- 1:9

all_tiles <- unique(df_polygons$fov)

for(i in all_tiles){

 gg <- df_final %>% filter(fov==i) %>% ggplot(aes(x = x, y = y, width=100)) +
  geom_tile(aes(fill = Cluster, group = id)) + theme_bw() + ggtitle(paste0("FOV=", i)) +
   scale_fill_manual(values=cols)
 
 ggsave(gg, file=paste0("../figures/5clusters_tiling_FOV_500_", i, ".jpg"), height=5, width=5, bg="white")
 
}
```

# Identify cell types in each cluster

```{r}
all_clusters <- cell_to_tile(df_final, sce, 100, cores=10, prob=FALSE)
all_clusters <-  unlist(all_clusters)

sce$Tile <- all_clusters[match(sce$Barcode, sapply(names(all_clusters), 
            function(x) strsplit(x, ".", fixed=TRUE)[[1]][2]))]
```

```{r}
res <- table(sce$Tile, sce$Anno)
res_enrich <- res/chisq.test(res)$expected

pheatmap(res_enrich, color = c(rev(colorRampPalette(brewer.pal(9, "Blues"))(10)), 
                               colorRampPalette(brewer.pal(9, "YlOrRd"))(9)), 
         breaks=c(seq(0, 1, 0.1), seq(2, 5, 1), seq(6, 10, 1)))
```

# Find genes expressed in tiles

```{r, fig.height=4, fig.width=3}
markers <- findMarkers(sce, groups=sce$Tile, test.type="t", pval.type="all")
markers_names_1 <- lapply(markers, function(x) rownames(x)[1:5])
markers_names_1 <- lapply(1:length(markers_names_1), function(x) setdiff(markers_names_1[[x]], 
    unlist(markers_names_1[-x])))
markers_names_1 <- unlist(markers_names_1)
 

agg_sce <- aggregateAcrossCells(sce, sce$Tile, subset.row=markers_names_1, use.assay.type="logcounts", statistics="mean")
sum_sce <- summarizeAssayByGroup(sce, sce$Tile, statistics="prop.detected",  subset.row=markers_names_1)

agg_sce <- reshape2::melt(agg_sce@assays@data$logcounts)
sum_sce <- reshape2::melt(sum_sce@assays@data$prop.detected)
colnames(agg_sce) <- c("Gene", "Tile", "Expression")
agg_sce$Tile <- factor(agg_sce$Tile, levels=1:9)
agg_sce$Gene <- factor(agg_sce$Gene, levels=c(markers_names_1))
agg_sce$Proportion <- sum_sce$value


gg1 <- ggplot(agg_sce, aes(y=Gene, x=Tile)) + geom_point(aes(size=Proportion, colour=Expression)) + 
  scale_colour_viridis_c() + theme_bw()

gg1

ggsave(gg1, file="../figures/Genes_tiles.jpg", height=7, width = 4, bg="white")
```

# Save files

```{r}
saveRDS(sce, file="../data/processed/sce_jj98_tiles.rds")

sce <- readRDS("../data/processed/sce_jj98_tiles.rds")
anno <- readRDS("../data/processed/annotation_all.rds")
sce$Anno <- anno$Annotation


res <- table(sce$Tile, sce$Anno)
res_enrich <- res/chisq.test(res)$expected


pdf("../figures/enrichment_plot.pdf", width=7, height=5)
pheatmap(res_enrich, color = c(rev(colorRampPalette(brewer.pal(9, "Blues"))(10)), 
                               colorRampPalette(brewer.pal(9, "YlOrRd"))(9)), 
         breaks=c(seq(0, 1, 0.1), seq(2, 5, 1), seq(6, 10, 1)))
dev.off()

```
